Bump the CAPI-release
As CVE-2017-8037 is pointing to the incomplete fix for the CC API access
to CC VM contents, we need to upgrade out CF-release to version v270 or
the CAPI-release to version 1.38.0.

We went with the CAPI upgrade, as it seems to be a smaller change as
oppose to upgrading the whole platform at the time, which we leave for
the future work.