api-avail: Silence apt-get output in task
The `apt-get` output from installing `awscli` in the API availability task
makes it hard to find the summary of the test. Reduce the output by:

- using `-qq` to make `apt-get update` super quiet
- redirecting STDOUT to `/dev/null` for `apt-get install`

The `-qq` flag didn't work for `apt-get install` because `dpkg` still writes
to STDOUT. I tried `-o=Dpkg::Use-Pty=0` too but that didn't help. We'll
still get errors from STDERR, so I think this is OK.