api-avail: Suppress cf push output from test
The output from pushing an app makes it hard to see the summary from the
test run. Suppress it by no longer running the tests in verbose mode, so
that `GinkgoWriter` is only flushed to STDOUT when a test fails.

Logging is sent to `os.Stdout` so that we can still see the final reporting
from successful tests.

We have some prior art for this in 8c05e87153a042cea4645275f6a55dd4c93b0b4f