Buildpack Release Notes Generator rebuilt
When upgrading Cloud Foundry we obtain a combined buildpack changelog to give our tenants. The existing shell script gave incorrect and inconsistent output. It wasn't reasonable to maintain and we rebuilt it in Python with many improvements.

The new script:

 - Finds submodules matching src/*-buildpack-release which have changed between the two releases.

 - Obtains a list of the interceding versions of the buildpack by comparing the new and old releases/.../index.yml.

 - Generates a summary markdown file.

 - Generates a detailed summary markdown file, containing buildpack changenotes fetched from Github's API.

Newly added and updated buildpacks are handled automatically, but we can't easily detect renamed or removed buildpacks. The script should not crash in such situations. We warn to check manually for them but neither have yet occurred in cf-release.