Order create-cloudfoundry pipeline first in concourse.
This makes it the default pipeline displayed when you go to the
concourse URL.

The fly command doesn't presently have a command to do this, so this
includes a ruby script to interact with the API using credentials read
from the .flyrc file. I've opened an issue[1] to request a command be
added to fly.

[1]https://github.com/concourse/concourse/issues/1561