Enable Redis and add acceptance tests
Adds the "redis" service and a "tiny" plan to the compose-broker
catalog.json.

Adds a redis endpoint to the healthcheck app and acceptance tests that
deploy and interact with the healthcheck app to ensure redis services
can be provisioned and bound to successfully.

The redigo redis client library was vendored for the healthcheck app.