Change to image_resource
A while ago, we've meen migrating from using the image to the new
image_resource instead. This one particular case, was remaining due to
the issues with `make <env> bosh-cli`. These problems are no longer
applicable and the change is required for the use with a new version of
concourse.