Rework splay task
We've hit some problems with the following scripts while upgrading our
concourse.

At first, task `inputs` needed to be nested under `config`. However
after we did that we got a deprecation warning that `config` cannot be
used with `file`. We weren't able to move the `inputs` to the task
because the task is used in two places, one which uses the input and one
which doesn't.