Refine creation of api-availability test app.
Give it a fixed name. This avoids stale app instances being left behind
in the persistent healthchecks space.

Reduce memory usage. By default this allocates 1G, which is unnecessary
as it typically uses less than 5M.

Specify the buildpack to speed up the staging process as it doesn't
need to detect which one to use.

Delete the route when deleting the app. This prevents stale routes being
left behind in the space. Prior to the change to use a fixed app name,
this would have eventually let to the quota for routes being exhausted.