Ensure ServerName is set in mysql TLS config in healthcheck app
The healthcheck app was updated to use a more recent Go version and the
latest compatible depenencies. The newer Go standard library is stricter
about ensuring that the ServerName field is correctly set for TLS.

The current version of the mysql library does not set the required
`ServerName` when using a connection string with 'tls=true'. A fix has
been already been merged and will be released in the next version (v1.4)
of the go-sql-driver/mysql lib.

See https://github.com/go-sql-driver/mysql/pull/564

To workaround the problem in the meantime, we register a "custom" TLS
config which correctly sets the ServerName.