Remove deprecated Compose secrets key
This key was the old name for this credential. When the name was changed
it was left in to preserve the function of existing deployment pipelines
that may rely on the old value. Plenty of time has passed and all environments
should now have the updated key in their Compose secrets file.