Deploy Compose Scraper according to ENABLE_DATADOG
The sole purpose of the Compose Scraper is reporting metrics, at
present to Datadog. The script is Datadog-specific and thus it makes
sense to gate the scraper behind `ENABLE_DATADOG` instead of a
dedicated flag.

The one downside here is that enabling Datadog requires either valid
Compose credentials or (if providing invented credentials) will cause
a failed login to Compose every 10 minutes.