Increase Elasticsearch disk size
Elasticsearch disk usage is rising in prod with increased usage and we
expect a small but significant jump in usage due to the increased log
volume from enabling the gorouter logs.

We are doubling the capacity for prod.

To prevent a change getting all the way to production without being
tested by runs through staging pipeline we are also making a minor
increase to the disk sizes for the default/staging environments. This
will cause the disk resizes during deployment in other environments to
ensure it is updates successfully.