Drop "ok" compose monitor in favour of "no data"
Previously there was a monitor configured that would notify if the compose scraper
did not report back with an "ok" health check. The health was only being
reported when the scraper started up and so was not very useful.

This has been dropped in favour of notifying when "no data" has been
returned for any of the other metrics that are being monitored (cluster
memory and disk usage).