Lower CHECK_INTERVAL_SECONDS for compose scraper
At 600s (10min) there were some inconsistancies with the stats reported
in datadog (some of the metrics appeared out of sync with the 10min
window, or would be missing). It's not totally clear why this was the
case since the scraper log was saying that it was reporting data at
regular intervals.

Lowering the interval to 300s (5mins) ensured more consitant reporting
and should still be far from any kind of rate limiting.