Fix upload-compose-secrets for staging and prod
This stopped working with the following error:

    ➜  paas-cf git:(master) ✗ make staging upload-compose-secrets
    Error: compose/staging/access_token is not in the password store.
    make: *** [upload-compose-secrets] Error 1

Because we have:

1. Moved the `compose/{staging,prod}/api_token` to
alphagov/paas-credentials-high and left `compose/dev/api_token` in
alphagov/paas-credentials.
2. Added `compose/billing/{email_address,password}` to
alphagov/paas-credentials which is used for all environments.

Fix this by setting an additional environment variable for the high
credential store location in the `staging` and `prod` make targets, and
using that to get the `api_token` if it's present.

This assume that you have the high credential store checked out in
`~/.paas-pass-high` which is the path also referenced by the
`manually_upload_certs` make target. We may look at refactoring or
standardising this in the future.