Use -startTime/-finishTime for fake start/stop
In the previous commit we added `-startTime` and `-finishTime` flags so that
we no longer needed to edit the input data to contain a single calendar
month.

This worked for instances where we saw the start and stop event within the
calendar month but it generated incorrect results for instances where we
didn't see the start or stop event. This is because the `continue` condition
happened after the `firstEvent` and `lastEvent` variables were populated, so
they were assumed to have been started before `-startTime` in the previous
month.

Fix this by re-using `startTime` and `finishTime` instead. I've renamed the
variable inside `processApp()` and `processEvent()` so that it's not
confused with `startTime` in a different context.