usage_reporting: Explain manually merging services
Explain that you need to manually merge the real and fake service usage
events. This is now the ugliest part of the process, now that we have more
organisations that are creating and destroying service instances. It would
be possible to combine the `generate_service_event` functionality with
`process_events` so that it only generates fake events for service GUIDs
that are currently running but don't appear in the event dump.