Modfiy GOPATH for acceptance-test runner
Go does not like symlinks in GOPATH and we were seeing errors/warnings
since upgrading to go1.9 for compiling the latest acceptance tests.

Rather than symlinking the cf-release resource into the /go dir we can
create a second GOPATH inside the build path (pwd) and mount the
resource directly in the correct place (to avoid any expensive unnecasary moving
or copying of the resources around in the container).

We also remove the symlink for the output dir and mount it directly at
/tmp/artifacts