Add cf_subshell_scoped_login helper script.
This is useful when you want to login to one of the persistent
environments for a specific task without that affecting other terminal
sessions.

For context, the CF CLI keeps information about the currently targeted
API endpoint along with auth tokens etc in $CF_HOME/.cf/config.json. If
CF_HOME is unset, it detaults to $HOME.

This script therefore creates a temporary directory and sets CF_HOME to
that before dropping into a subshell. An exit trap cleans all this up.