Add uaa_lock_user.rb script
This will replace the rather cumbersome process for locking (or unlocking)
users that's currently described in the team manual:

- https://government-paas-team-manual.readthedocs.io/en/latest/guides/CF_UAA_user_management/#locking-a-user-account

I've mocked just enough of the UAA responses to make it work and verify that
we're sending PATCH requests when the state needs to change.

Per the original script, I've left TARGET and TOKEN as environments, to call
out the fact that they are mandatory and to help avoid you storing tokens in
your shell history.