Add RDS Instance management
We believe it will decrease the costs for the develoment environments to
turn off the RDS instances over the night along with the rest of
autodelete-cloudfoundry pipeline.

These will start back up in the morning, with the use of
deployment-kick-off pipeline.

We've prepared a script which could be also used locally if needed in
order to not duplicate the same functionality before start/stop call.