Fix cc_uploader server cert
The cc_uploader service was using the cloudcontroller server cert. When
we enabled TLS for connections from rep to cc_uploader the connections
failed with the error:

x509: certificate is valid for cloud-controller-ng.service.cf.internal, not cc-uploader.service.cf.internal

This caused the staging process to fail when attempting to upload the
resulting droplet.

We fix this by generating a dedicated server cert for cc_uploader with
the correct name in the SAN. This follows what's documented upstream[1].

[1]https://github.com/cloudfoundry/capi-release/blob/1.42.0/docs/tls-configuration.md#generating-the-cc-uploader-certificates