Workaround for bosh-cli timeout
The current bosh-cli make target works by hijacking a concourse
container that has recently run (and succeeded) to use as a temporary
shell to access bosh.

Some recent concourse updates changed the behaviour of how long and when
old containers are garbage collected, and where it used to be the case
that if you were logged in to a hijacked container the container would
be protected from garbage collection (or at least stay around for a very
long time), it is now the case that containers are cleaned up after
around 5mins and we get kicked out of our bosh-cli shell.... which is
frustrating.

To workaround this issue the following changes are made:

* the run-bosh-cli task will now sleep for as long as a pseudo-terminal
and only 'succeed' after 5mins of inactivity
slave is detected (ie someone has a terminal wired up to the container)
* the script to launch the job will wait until the container is running,
then connect to it
* the script to launch to shell will abort the job on EXIT to clean up
faster