Don't delete healthcheck app when deleting DB
This task runs early in the autodelete-cloudfoundry pipeline, and
previously would delete the healthcheck app so that it's database could
be deleted. If a job later in the pipeline failed, the system would be
left in a state where the app had been deleted, but the flag file in the
state bucket still indicated it was deployed. This then caused problems
in the next deployment as the availability tests would run and fail.

There's no need to delete the app in this task, and it's non-obvious
from the name that it will be. This therefore updates the task to merely
unbind the app from the database before deleting the database.