bosh-cli download manifest from bosh
This is currently using the manifest from the state bucket. There have
been situations where this manifest doesn't match what's currently
deployed to bosh. In this case it can cause problems when interacting
with bosh because it complains that a deployment hasn't completed etc.

Switching this task to instead download the manifest from the bosh
director ensures that the manifest we're using matches Bosh's view of
the world.