Add instance detailing script
As part of #147674123 we would like to have an easy way of establishing
the belongings of an individual instance.

Given we obtain an instance ID from either DataDog, Kibana or AWS
Console, we can run this script to learn which org and space the
instance belong to along with it's name and list of org managers.