shellcheck pipeline substitutions for double-quotes.
pipecleaner.py uses a regex to substitute all fly interpolations in the
pipelien with a dummy value. This causes shellcheck to think they're
safe, and not warn about double-quoting etc for strings containing them.

By extending the substituted value to include `$(date)` the strings no
longer look safe to shellcheck and it will therefore warn about double
quoting etc.

This includes fixing occurrences of this in the pipelines following the
pipecleaner change.