shellcheck concourse pipelines for unused variables.
We want to catch unused variables in our pipelines as these are a code
smell.

Fixing the resulting errors has required refactoring how we set the
config variables to avoid the loop with the `eval` in it, and instead
use cat with a heredoc.

I've also had to update the reged pipecleaner was using to replace
interpolation variables as the previous one was being too agressive and
was also replacing bash $((...)) arithmetic expressions. The regexp now
matches the one that fly uses.