Upgrade stemcell to 3445.16
This supersedes the upgrade to 3445.15 in 961e427d because that stemcell
doesn't have a "light" version published. Meaning that:

- the same URL redirects to the "fat" version instead
- it takes longer to download/upload
- it possibly requires more time recompiling packages when deploying VMs,
  but haven't haven't confirmed that
- it's not usable with bosh-init from a container

In order to be consistent with previous deployments and the stemcell that we
have to use in paas-bootstrap we're going to skip .15 in production and go
straight to .16 - the only difference in the contents are Ubuntu package
updates.