Amend ELB latency monitors
We realised that the CC ELB will create a lot of noise in our production
environment. We decided to amend the thresholds as well as the window to
for the metrics to be monitored.

I've updated the window on other monitors too and took the thought to
change the latency check from "both" to "above" as we probably don't
care if it took less time than usual to serve the response.