Remove Nginx from in front of UAA
Having another reverse proxy in front of UAA was causing some requests to fail
the api-availability-tests when a deployment restarted the UAA VMs:

    2017-11-07 09:52:26.23 UTC - 2017-11-07 09:52:26.26 UTC Listing all apps in a space: Error getting token: oauth2: cannot fetch token: 502 Bad Gateway
    Response: <html>
    <head><title>502 Bad Gateway</title></head>
    <body bgcolor="white">
    <center><h1>502 Bad Gateway</h1></center>
    <hr><center>nginx/1.11.7</center>
    </body>
    </html>

We introduced this in 681659ea because we needed a way to block the Create
Account and Forgot Password endpoints. Since then the behaviour of UAA's
`login.self_service_links_enabled` property has changed - it used to only
remove the links from the UI, but it now also disables the endpoints:

- https://www.pivotaltracker.com/n/projects/997278/stories/141508943

nginx-release and the logsearch filters were only used for this job so they are
no longer needed.

With the removal of Nginx we no longer have access logs from the UAA VMs in
Logsearch. We do have access logs from gorouter, which will help us to see
traffic but not necessarily diagnose inter-VM problems. I've confirmed that
the `uaa.audit.remote_address` in UAA's audit logs still has the correct IP
from `X-Forwarded-For`.