Set UAA ELB drain timeout to 15s and gorouter drain wait to 30s
We enabled the UAA ELB connection draining but didn't set the timeout which is
300s by default. This meant that long-living connections (e.g. the API
availability tests probably used keep-alive connections) could still reach the
haproxy on an unhealthy instance after the gorouter was stopped after
drain_wait time. That's why we were getting 503 errors.

The drain_wait time has to be higher than the maximum time to set an instance
to unhealthy + the ELB draining timeout.

We set the ELB draining timeout to 15 seconds. For an unhealthy instance we
need maximum 12 seconds (2 checking intervals + health check timeout: 2 * 5 +
2), so we need the drain_wait higher than 15 + 12 seconds therefore we set it to 30
seconds.