Set drain_wait to 120s on gorouter
On further testing we still saw 503 errors and checking the AWS documentation
it turned out that the connection draining timeout is not used for unhealthy
instances.

From: http://docs.aws.amazon.com/elasticloadbalancing/latest/classic/config-conn-drain.html

> If an instance becomes unhealthy, the load balancer reports the instance state
as OutOfService. If there are in-flight requests made to the unhealthy
instance, they are completed. The maximum timeout limit does not apply to
connections to unhealthy instances.

By setting the drain_wait to a higher value it seems the number of errors will
decrease. Checking the http client in the go-cfclient there is no maximum limit
for keeping alive a persistent connection. Also it seems the ELB keeps the
connection alive between the unhealthy instance and the load balancer as long
as the client keeps the connection alive to the load balancer.