New paas-uaa-assets app to serve custom stylesheet
In a subsequent commit we are configuring UAA to use a custom location
for its stylesheet. This requires something to be serving a stylesheet
and for consistency with the rest of the platform we are serving this
stylesheet from a new paas-uaa-assets app.

This commit deploys the new paas-uaa-assets app, which to avoid
missing stylesheets between the UAA update and the app deploy must be
done before the UAA update. The latter commit will then perform the
UAA update once the app is deployed.

This also creates a new `assets` space in the `admin` org, as none of
the existing spaces were relevant and this app is something of a
one-off.