Restyle the 'Welcome to PaaS' email
This change adds a HTML version to our 'Welcome to PaaS' email which we send
when we create a new user or reset the password for an existing user.

I extracted the existing plain text template to a Ruby ERB template and saved
the new HTML template next to it. The HTML code would be hard to handle in the
script and this way the templates are easier to manage.