Use the new Go Router health check on the UAA and router ELBs
As we start to use the new health check we have to point the ELB health checks
to the /health path. During release the old health check endpoint will still
respond as expected as it's ignores the path.

The HAProxy upgrade is required to remove the now redundant `HTTP-Monitor/1.1`
User Agent-header and set up a dedicated backend configuration for healthcheck
requests.