Don't rotate CF database user passwords
We won't rotate the following secrets:
 * cf_db_api_password
 * cf_db_locket_password
 * cf_db_uaa_password

These passwords are updated in the cf-terraform job, but the app configurations
are only updated in the cf-deploy job. It can cause downtime if we change these
passwords for existing users. As the databases can't be accessed from outside
(only from our VPC) the risk is low if we don't rotate them.