Add custom redirection on user invite
When the user receives an email to complete their registration it will
ask them to provide their password. Once the password is set
successfully, user has accepted an invitation and will be redirected to
the UAA home page.

It isn't very useful page, especially if you're just starting off with
CF. We've created our own landing page we'd like to be providing to our
users upon registration.