val_from_yaml.rb: Fix array lookup by position
Rubocop autocorrect in 97a8a506 changed the order of a comparison which
broke `val_from_yaml.rb` when looking up array items by position. We noticed
this when `get-and-upload-stemcells` in the `create-cloudfoundry` pipeline
no longer uploaded the stemcell so a fresh deployment failed with:

    Error 50003: Stemcell 'bosh-aws-xen-hvm-ubuntu-trusty-go_agent/3468.5' doesn't exist

The use of `===` meant that the order of the comparison arguments were
significant. When regex is the first argument it compares the other value to
the regex, but when a string is the first argument it also asserts that the
other argument is also a `String`.

- https://apidock.com/ruby/Regexp/%3D%3D%3D
- https://apidock.com/ruby/String/%3D%3D%3D

Fix this by explicitly using `=~` to compare to a regex. I've added a test
to prevent this happening again.