Deploy the ElastiCache Redis broker to Cloud Foundry
We introduce the new broker to offer ElastiCache Redis services to our tenants.
The new service will be called redis.

The broker will be installed to two small VMs which will be load balanced with
an ELB. The brokers will be able to authenticate to the AWS API with an
instance profile.
All the Redis replication groups will be in the same Cache subnet group and
will have the same security group.
The cells will have a new security group to be able to connect to the Redis
instances.

For now we only introduce one plan which is called tiny and will be essentially
the trial plan.