Overcommit cell memory by 66%
Minimising excess running costs means reducing our unused capacity.
This commit will allow us to make use of previously wasted memory
capacity on our cells.

App instances reserve a given amount of memory, but we're seeing that
they generally reserve a lot more memory than they utilise. At present
apps on average only use 34% of their reserved memory. We've never
seen this higher than 54%, which was a long time ago.

To start using this excess capacity, we'll be quite conservative and
say that the number will never go above 60%. If apps only use 60% of
their reserved memory, then we can fit more apps into the "reserved"
memory that they don't actually use.

We want to find a capacity that `rep` should be advertising. Such a
capacity, when fully used by apps that use 60% of their reservation,
would result in 100% of the cell memory being utilised by apps. We can
formulate this as an equation:

  c * 0.6 = 1.0
  c = 1.0 / 0.6 = 1.666...

We round this to 1.66 for simplicity. It means that we want to
overcommit our cells by 66% in order to use the expected 40+%
underutilisation.

The next question is how much that percentage overcommit is in terms
of MiB. Our cells run on AWS `r4.xlarge` instances. AWS advertise
these as having `30.5GiB = 31232MiB` of RAM. In practice we seem to
have a little less total memory:

* On our `r4.xlarge` cells, `free -m` lists `30662` MiB
* The `cf.rep.CapacityTotalMemory` metric shows that `rep` is
  advertising `30662MiB` as a cell's total capacity on all staging and
  production cells

We overcommit this `30662MiB` by the 66% calculated above.

We have a monitor set to alert us if actual usage on the cells exceeds
a threshold, so if apps on the platform start using more of their
reservation we would expect to be notified by the monitor.