Reduce load on API during availability tests
We want to avoid putting the API and UAA under unnecessary load during
availability tests. We only care about the API being continually available,
this is not a form of load testing.

The tasks performed against the API are not simple, single-request tasks,
so we can't use a library such as Vegeta to perform the rate limiting.
For simplicity, the rate at which tasks are added to the queue has been
throttled.

Under the new configuration we will not need as many workers, as only
two tasks per second will be added to the queue.