Make UAA client configurations override
According to the UAA docs[1] this property allows the configuration
in the manifest to override client configurations saved if you have
a persistent database. We set this because we always want our manifest
to be the source of truth for configuration.

Also, we remove the `id` fields because they are redundant. The UAA
release uses the `properties.uaa.clients.*` key as the ID.

[1] https://github.com/cloudfoundry/uaa/blob/391163ebad397b8f3eb5298aa01412dd94c9a176/docs/Sysadmin-Guide.rst#clients