Share the System DNS Zone Name with govuk-paas org
We want to map subdomains of `cloud.service.gov.uk` to the product
page and the tech docs. These apps are deployed by CI into the
`govuk-paas` organisation. Mapping those subdomains therefore requires
the `govuk-paas` organisation to be able to manage routes of that
domain, which we name the `$SYSTEM_DNS_ZONE_NAME`.