Redirect to the next steps page when users log in
We'd like users to go to https://www.cloud.service.gov.uk/next-steps
when they log in. We previously implemented this for when users accept
an invitation, but discovered in user research that there was no such
redirect when users log in.

The `login.links.global.homeRedirect` parameter [1] appeared suited,
but we found the redirect did not take effect. We've found that the
`login.links.homeRedirect` (note the missing `global`) works.

[1] https://github.com/cloudfoundry/uaa-release/blob/b8b2caae7d705ecde6112b614627ea1abb2ace24/jobs/uaa/spec#L293-L299