Add a gauge for monitoring TLS cert validity
We want to be warned if any public facing certificates are near expiry
or become invalid.

This adds a new "gauge" that periodically fetches the certificates
returned for a CSV set of domains (configured via TLS_DOMAINS env var),
and then reports back the number of days before each cert expires.

If there is a problem with the certificates (such as it is incorrectly
configured for the given domain, is invalid or is from an untrusted CA
then the reported value will fall to zero to denote invalid/expired.

The metric is labelled: `tls.certificates.validity` and the unit is
"days"

So you could configure alerts at `< 10 days` to give time to fix a cert
that will expire soon or `= 0` to alert on problems that may be
impacting tennants immediately.