Delete `documentation` UAA client in `cf-deploy`
We have removed the `documentation` UAA client from the manifest, but
this does not remove it from UAA. This commit adds a task to the
`cf-deploy` job of the `create-cloudfoundry` pipeline, to perform the
deletion of the `documentation` client. We anticipate that this commit
will be reverted after it has been deployed against our persistent
environments.