Deploy pass-metrics based on ENABLE_DATADOG flag
Previously we used an ENABLE_METRICS flag to decide when the
paas-metrics app should be deployed. The reasoning was that more metrics
(and more tagged metrics) costs more in datadog.

However it's not great to not deploy this in staging and we do not think
that the larger number of metrics should affect things too much due to
staging having much less running.

We also want to monitor the TLS certs in staging using paas-metrics.

So we have removed the ENABLE_METRICS flag and paas-metrics will always
be deployed based on the ENABLE_DATADOG flag.