Include the apex apps domain in TLS checks
The apex apps domain is used in some scenarios (such as cloudfront
origins). Since it's easy to missconfigure a TLS cert to miss this case
(since not all wildcard cert have an alt name for the apex) we want to
ensure we are checking validity of this directly.

The list was also getting quite long so it was refactored into an array
in the deployment for readability.