Update ELB log bucket policy
This commit updates the policy in place for ELB log writing access to S3
in order to remove the false positives recently seen in Trusted Advisor.

I have elected to use the option that AWS indicated in their last
response to the support ticket to solve this issue as it is a more
straightforward approach and keeps the flexibility in place to add
additional ELB's using our exsisting approaches.