terraform: Require 0.11.1 for manual runs
We're upgrading Terraform everywhere and 0.11.1 is the latest stable.

I've specified a specific version rather than a pessimistic version because
we don't have much control over which versions of Terraform people have
installed locally on their laptops and if someone uses a newer versions then
the state file will force everyone else to upgrade.

This could be improved in future by using the container that we use in the
pipeline.