terraform/cf-certs: Upgrade to Terraform 0.11.1
We need to write a temporary file to indicate that we want to use S3 as the
remote state backend. This can't be present in the repo because the same TF
codebase is also used in the pipeline where we use a local state file that
is downloaded/uploaded as a Concourse resource.