Delete admin users from CF, not UAA
We want to use the CF API to delete users, not UAA. This is because
deleting them from UAA leaves an orphaned record in the Cloud Controller
database.

The UAA token being used was one from the UAA admin client, using the client
credentials grant. This token is not scoped to perform admin operations
in Cloud Foundry, so we have to get a user's token via the `cf` UAA client
and our CF admin user's name and password.