Configure UAA with SMTP details for AWS SES
SMTP credentials have to be derived from a user's access keys, so we
create the user in the pipeline and assign it to a group that has permission
to send raw SES emails. Terraform exports an attribute with the derived
SMTP password, so we don't have to script it ourselves.