Enable password resets and block account creation
* enable self-service links for resetting passwords and creating
  accounts.
* create an app that acts as a holding page for all requests to
  create an account.
* create domains for `{login,uaa}.system_domain`, create routes for the
  account creation endpoints, and map them to the holding page

We decided not to use the blue-green-deploy plugin for deploying the holding
page. This has a bug which results in it registering the `login.system_domain`
(with no path) every time you deploy. This means UAA traffic incorrectly
gets sent to the holding page[1].

The Nginx config for the holding page app has a separate listener on port 8989
because Nginx doesn't allow serving static content for `POST` requests.

We hardcoded the `homeRedirect` URL to be consistent with the user invitation
client, and because the product page application has a custom domain in prod.

[1] https://github.com/bluemixgaragelondon/cf-blue-green-deploy/issues/57