Combine db removal scripts.
These are currently very similar. The only difference between them
(other than that database etc they operate on) is that the billing one
deletes the app instead of unbinding it before deleting the database. I
believe the only reason for this is that this database was added
(c8dcd58) in parallel with the healthehck script being updated to unbind
(00e4cb4)