Add credential rotation for SES SMTP user
We want to avoid long-lived AWS keys for the SES SMTP user. The job
for removing the key from the Terraform state file is safe to do at
any time, because only the state file is changed, the key will still
exist. Because of this we can run the job monthly on a timer. A
subsequent run of the pipeline will result in the `cf-terraform`
job creating a new access key and any unused keys will be deleted.

Rotation of credentials automatically is our ultimate goal. We haven't
established this as a pattern yet, but we took the opportunity to do it
here because it was easy to implement.